+++ http4k-client-okhttp/src/test/kotlin/org/http4k/client/OkHttpAsyncTest.kt
@@ -1,0 +1,5 @@
+package org.http4k.client
+
+import org.http4k.server.SunHttp
+
+class OkHttpAsyncTest : Http4kAsyncClientContract({ SunHttp(it) }, OkHttp(), OkHttp(timeout))