+++ http4k-core/src/test/kotlin/org/http4k/client/Http4kClientContract.kt
@@ -14,63 +14,19 @@ import org.http4k.core.Method.POST
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
-import org.http4k.core.Status.Companion.BAD_REQUEST
-import org.http4k.core.Status.Companion.FOUND
 import org.http4k.core.Status.Companion.OK
 import org.http4k.core.then
 import org.http4k.filter.ClientFilters
 import org.http4k.filter.ServerFilters
-import org.http4k.routing.bind
-import org.http4k.routing.path
-import org.http4k.routing.routes
-import org.http4k.server.Http4kServer
 import org.http4k.server.ServerConfig
 import org.http4k.server.SunHttp
 import org.http4k.server.asServer
-import org.http4k.util.RetryRule
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import java.nio.ByteBuffer
-import java.util.Arrays
-import java.util.Random
 
-
-abstract class Http4kClientContract<out T : HttpHandler>(private val serverConfig: (Int) -> ServerConfig,
-                                                         val client: T,
-                                                         val timeoutClient: T) {
-    @Rule
-    @JvmField
-    var retryRule = RetryRule(5)
-
-    private var server: Http4kServer? = null
-
-    val port = Random().nextInt(1000) + 8000
-
-    @Before
-    fun before() {
-        val defaultHandler = { request: Request ->
-            Response(OK)
-                    .header("uri", request.uri.toString())
-                    .header("header", request.header("header"))
-                    .header("query", request.query("query"))
-                    .body(request.body)
-        }
-        server = routes("/someUri" bind POST to defaultHandler,
-            "/empty" bind GET to { _: Request -> Response(OK).body("") },
-            "/redirect" bind GET to { _: Request -> Response(FOUND).header("Location", "/someUri").body("") },
-            "/stream" bind GET to { _: Request -> Response(OK).body("stream".byteInputStream()) },
-            "/delay/{millis}" bind GET to { r: Request ->
-                Thread.sleep(r.path("millis")!!.toLong())
-                Response(OK) },
-            "/echo" bind POST to { request: Request -> Response(OK).body(request.bodyString()) },
-            "/check-image" bind POST to { request: Request ->
-                if (Arrays.equals(testImageBytes(), request.body.payload.array()))
-                    Response(OK) else Response(BAD_REQUEST.description("Image content does not match"))
-            })
-            .asServer(serverConfig(port)).start()
-    }
+abstract class Http4kClientContract(serverConfig: (Int) -> ServerConfig,
+                                    val client: HttpHandler,
+                                    private val timeoutClient: HttpHandler) : AbstractHttp4kClientContract(serverConfig) {
 
     @Test
     fun `can forward response body to another request`() {
@@ -92,16 +48,11 @@ abstract class Http4kClientContract<out T : HttpHandler>(private val serverConfi
         asServer.stop()
     }
 
-    @After
-    fun after() {
-        server?.stop()
-    }
-
     @Test
     fun `can make call`() {
         val response = client(Request(POST, "http://localhost:$port/someUri")
-                .query("query", "123")
-                .header("header", "value").body("body"))
+            .query("query", "123")
+            .header("header", "value").body("body"))
 
         assertThat(response.status, equalTo(OK))
         assertThat(response.header("uri"), equalTo("/someUri?query=123"))
@@ -189,6 +140,4 @@ abstract class Http4kClientContract<out T : HttpHandler>(private val serverConfi
 
         assertThat(response.status, equalTo(Status.CLIENT_TIMEOUT))
     }
-
-    private fun testImageBytes() = this::class.java.getResourceAsStream("/test.png").readBytes()
 }